refactor backtest into files
